Interpreter: Don't locally execute the args of a lambda
When defining higher-order functions that take lambdas, like map, it's
essential that we don't execute the arguments of map within the local
environment of the lambda, as we then lose the parameters that we have
previously bound.  As a simple example, consider

    (define f (g x)
      (g x))

Calling (f (lambda (y) (+ y 1)) 1) would bind the (evaluated!) lambda to
g and 1 to x, but the lambda's environment (being created when it was
first evaluated) doesn't know about x at all and would throw an error.